refactor: 함수 분리 및 필요없는 코드, 주석  삭제
